Tasks: don't advance task RNG on task spawn
Previously we had this unfortunate behavior:

    julia> Random.seed!(123)
    TaskLocalRNG()

    julia> randn()
    -0.6457306721039767

    julia> Random.seed!(123)
    TaskLocalRNG()

    julia> fetch(@async nothing)

    julia> randn()
    0.4922456865251828

In other words: the mere act of spawning a child task affects the parent
task's RNG (by advancing it four times). This PR preserves the desirable
parts of the previous situation: when seeded, the parent and child RNG
streams are reproducible. Moreover, it fixes the undesirable behavior:

    julia> Random.seed!(123)
    TaskLocalRNG()

    julia> randn()
    -0.6457306721039767

    julia> Random.seed!(123)
    TaskLocalRNG()

    julia> fetch(@async nothing)

    julia> randn()
    -0.6457306721039767

In other words: the parent RNG is unaffected by spawning a child.

The design is based on the SplitMix [1] and DotMix [2] RNGs, but with
some simplifications based on observing that we only need binary
pedigree coefficients: when a task forks a child, zero is appended to
its pedigree, and the child's pedigree is the same with a one in the
last place. Thus all the pedigree coefficients are binary. How does this
help matters? In the of collision resistance in the DotMix paper,
working in a prime modulus is only necessary to guarantee that the
difference between pedigree coordinates has a multiplicative inverse. If
the coefficients are binary, then the difference is always 1, which
means we can work in any modulus, including Z/2^64, which lets us use
native integer arithmetic.

Similar to SplitMix, instead of explicitly storing pedigree coordinates,
we store each task's dot product and derive each child task's dot
product by adding to the parent dot product, the random weight
coefficeint for the current tree depth. No multiplication is required,
we just add the random weight for the current tree depth to the parent's
dot product to get the child's dot product.

Pseudorandom weights for the SplitMix dot product are generated using an
internal PCG RNG, specifically the PCG-RXS-M-XS variant. We chose this
RNG because we need something small and fast but unlikely to have
artifacts which might sabbotage the collision resistance of SplitMix.
Since PCG-RXS-M-XS passes BigCrush with only 36 bits of state and we use
64 bits of state, it fits the bill quite well. The major caveat of this
RNG is that it produces each value once, which makes it insecure, but
this is actually beneficial here: we don't want any repeated weights. We
also don't care that this RNG is insecure and invertible since it's only
used internally to seed forked child tasks. Rather than the classic LCG
multiplier inherited from Knuth, we use the best full width multiplier
found by Steele & Vigna [3] searching for high spectrum constants.

We reuse the PCG-RXS-M-XS output function for mixing the bits of the
SplitMix dot product, instead of the MurmurHash3 mixing function that it
normally uses since we're already using for the internal PCG generator.
The choice of mixing output function for SplitMix is essentially
arbitrary and only serves to cascade bit differences in dot products.

[1]: https://gee.cs.oswego.edu/dl/papers/oopsla14.pdf

[2]: http://supertech.csail.mit.edu/papers/dprng.pdf

[3]: https://arxiv.org/pdf/2001.05304.pdf